Extract generateToken out to seperate file and write simple unit test for it for good measure
