+++ api/server.spec.js
@@ -2,7 +2,6 @@ const request = require('supertest');
 
 const server = require('./server.js');
 const db = require('../database/dbConfig.js');
-const { intersect } = require('../database/dbConfig.js');
 
 
 describe('auth-router.js', () => {
